style: fold debate r2 LOW nits (testlib rename, docs, testStartUnixUs)
N1: createEndpoint.forTest → createEndpoint.testlib (house idiom).
N2: restore file-header docs above imports in restartLocal/padiBinding tests.
N3: GateIdentityRead startUnixUs docs match exact-v2 rule.
N4: recycleForeignGate uses testStartUnixUs instead of pid * 1_000 magic.
