+++ packages/surface-daemon-supervisor/src/squatterRecovery.test.ts
@@ -27,7 +27,7 @@ import {
   type EndpointStatus,
   isSocketSquatterForeignError,
 } from "./endpoint.ts";
-import { createEndpointForTest as createEndpoint } from "./createEndpoint.forTest.ts";
+import { createEndpointForTest as createEndpoint } from "./createEndpoint.testlib.ts";
 
 import { endpointPrivate } from "./endpoint.private.ts";
